+++ ui/src/app/pages/workspace/workspace-wrapper.tsx
@@ -3,7 +3,7 @@ import {Spinner} from 'app/components/spinners';
 import {WorkspaceNavBar} from 'app/pages/workspace/workspace-nav-bar';
 import {withCurrentWorkspace} from 'app/utils';
 import {routeDataStore, useStore} from 'app/utils/stores';
-import {WorkspaceRoutes} from 'app/workspace-app-routing';
+import {WorkspaceRoutes} from 'app/routing/workspace-app-routing';
 import * as fp from 'lodash/fp';
 import * as React from 'react';
 import {useEffect} from 'react';
