+++ ui/src/app/routing/workspace-app-routing.tsx
@@ -5,23 +5,23 @@ import * as fp from 'lodash/fp';
 import * as React from 'react';
 import {useEffect} from 'react';
 import {Redirect, Switch, useRouteMatch} from 'react-router-dom';
-import {NOTEBOOK_PAGE_KEY} from './components/help-sidebar';
-import {InteractiveNotebook} from './pages/analysis/interactive-notebook';
-import {NotebookList} from './pages/analysis/notebook-list';
-import {NotebookRedirect} from './pages/analysis/notebook-redirect';
-import {CohortReview} from './pages/data/cohort-review/cohort-review';
-import {DetailPage} from './pages/data/cohort-review/detail-page';
-import {QueryReport} from './pages/data/cohort-review/query-report.component';
-import {ParticipantsTable} from './pages/data/cohort-review/table-page';
-import {CohortActions} from './pages/data/cohort/cohort-actions';
-import {ConceptHomepage} from './pages/data/concept/concept-homepage';
-import {ConceptSearch} from './pages/data/concept/concept-search';
-import {ConceptSetActions} from './pages/data/concept/concept-set-actions';
-import {DataComponent} from './pages/data/data-component';
-import {DatasetPage} from './pages/data/data-set/dataset-page';
-import {WorkspaceAbout} from './pages/workspace/workspace-about';
-import {WorkspaceEdit, WorkspaceEditMode} from './pages/workspace/workspace-edit';
-import {BreadcrumbType} from './utils/navigation';
+import {NOTEBOOK_PAGE_KEY} from '../components/help-sidebar';
+import {InteractiveNotebook} from '../pages/analysis/interactive-notebook';
+import {NotebookList} from '../pages/analysis/notebook-list';
+import {NotebookRedirect} from '../pages/analysis/notebook-redirect';
+import {CohortReview} from '../pages/data/cohort-review/cohort-review';
+import {DetailPage} from '../pages/data/cohort-review/detail-page';
+import {QueryReport} from '../pages/data/cohort-review/query-report.component';
+import {ParticipantsTable} from '../pages/data/cohort-review/table-page';
+import {CohortActions} from '../pages/data/cohort/cohort-actions';
+import {ConceptHomepage} from '../pages/data/concept/concept-homepage';
+import {ConceptSearch} from '../pages/data/concept/concept-search';
+import {ConceptSetActions} from '../pages/data/concept/concept-set-actions';
+import {DataComponent} from '../pages/data/data-component';
+import {DatasetPage} from '../pages/data/data-set/dataset-page';
+import {WorkspaceAbout} from '../pages/workspace/workspace-about';
+import {WorkspaceEdit, WorkspaceEditMode} from '../pages/workspace/workspace-edit';
+import {BreadcrumbType} from '../utils/navigation';
 
 const CohortPagePage = fp.flow(withRouteData, withRoutingSpinner)(CohortPage);
 const CohortActionsPage = fp.flow(withRouteData, withRoutingSpinner)(CohortActions);
