+++ ui/src/main.ts
@@ -6,7 +6,7 @@ import * as ReactModal from 'react-modal';
 
 import {ErrorHandler} from 'app/components/error-handler';
 import {setupCustomValidators} from 'app/services/setup';
-import {AppRoutingComponent} from './app/app-routing';
+import {AppRoutingComponent} from './app/routing/app-routing';
 
 setupCustomValidators();
 ReactModal.defaultStyles.overlay = {};
